moved component specific css test to new folder
